% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNames.R
\name{cNames}
\alias{cNames}
\title{DF Naming Function}
\usage{
cNames(x)
}
\arguments{
\item{x}{A df}
}
\value{
renamed columns
}
\description{
This function allows naming of 5 column df to predetermined
names.
Required to implement maxfit() function.
}
\examples{
cNames(YI1)
cNames(CM)
}
\author{
Kristen Rodriguez
}

