% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropcol.R
\name{dropcol}
\alias{dropcol}
\title{Maxfit Naming Function}
\usage{
dropcol(x)
}
\arguments{
\item{x}{A df}
}
\value{
columns not dropped
}
\description{
This function allows dropping of column from the frame
can be modified for any/multiple columns
}
\examples{
nYI1(YI1)
}
\author{
Kristen Rodriguez
}

