% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapit.R
\name{mapit}
\alias{mapit}
\title{A Tree Maping Function}
\usage{
mapit(x)
}
\arguments{
\item{x}{A df}
}
\value{
A treemap
}
\description{
This function allows a treemap to be generated
}
\examples{
mapit(YI1)
}
\author{
Kristen Rodriguez
}

