% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxfit.R
\name{maxfit}
\alias{maxfit}
\title{Scatterplot Funtion}
\usage{
maxfit(x, ...)
}
\arguments{
\item{x}{A df}
}
\value{
The maximum fitness grouped by generation
}
\description{
This function allows creation of a scatterplot from a two column
data frame.
exact naming
}
\examples{
maxfit(YI1, Generation, Fitness)
maxfit(CM, Generation, Fitness)
}
\author{
Kristen Rodriguez
}

