% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnames.R
\name{maxnames}
\alias{maxnames}
\title{Maxfit Naming Function}
\usage{
maxnames(x)
}
\arguments{
\item{x}{A df}
}
\value{
renamed columns
}
\description{
This function allows optional naming of maxfit() output columns
to predetermined names
}
\examples{
maxnames(YI1)
maxnames(CM)
}
\author{
Kristen Rodriguez
}

