% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot Function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{x}{A df}
}
\value{
multiple plots arranged in grid
}
\description{
This function allows creation of multiple ggplot objects
displayed in a single plot
}
\examples{
multiplot(s.YI1, s.YI2, s.YI4, s.VD4, s.CM, s.DP1,cols=3)
}
\author{
Kristen Rodriguez
}

