% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot Funtion}
\usage{
scatterplot(x)
}
\arguments{
\item{x}{A df}
}
\value{
A scatterplot
}
\description{
This function allows creation of a scatterplot from a two column
data frame.
Specifically used for visualizing output of maxfit() function
}
\examples{
scatterplot(YI1)
scatterplot(CM)
}
\author{
Kristen Rodriguez
}

